+++ app/features/providers/github/commands/misc/tryParseDescription.ts
@@ -1,0 +1,10 @@
+import type { Maybe } from '~/types'
+import { marked } from 'marked'
+
+export const tryParseDescription = (description: string): Maybe<string> => {
+  try {
+    return marked(description)
+  } catch (e) {
+    return description
+  }
+}
