+++ package.json
@@ -55,10 +55,12 @@
     "@remix-run/react": "^1.12.0",
     "@remix-run/serve": "^1.12.0",
     "@remix-run/server-runtime": "^1.12.0",
+    "@types/marked": "^4.0.8",
     "@zentered/issue-forms-body-parser": "^2.1.1",
     "date-fns": "^2.29.3",
     "date-fns-tz": "^1.3.7",
     "graphql": "^16.6.0",
+    "marked": "^4.2.12",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "remix-image": "^1.4.0"
